Fix polyphonyControl.orc state UDO release stage issues
The `lastcycle` opcode is returning `1` when the note is released instead of waiting until the envelope's release stage completes.

This change fixes the issue by calling the envelope UDO before the polyphony control UDO.

This change also fixes an issue in the `AF_PolyphonyControl_audioProcessing` UDO that does not affect the combo-a1 instrument, but may affect instruments created in the future.